Implement the fast doubling Fibonacci strategy.
It is almost finished but need a multi precision multiplication implementation the one from numerical recipes is wrong.

This method works up to F(93) but I can not get the nr C code to work for larger integers - there are sign flipping bugs and lots of other bugs.  I have double checked and the code is correct according to the book.

I need to take another approach in the next commit.

last commit before removing nr code